Updated comments and MiniTimeManager class
Changed the comments for all the classes to reflect the process of the code, and I added MiniTimeManager to help manage the absolute/default cooldowns of the MiniTimes.
MiniTimes won't be holding the morale, max & min variables anymore. Finally, I added a small amount of empty variables and the emptyUpdate method